\name{npiv-package}
\alias{npiv-package}
\docType{package}
\title{
\packageTitle{npiv}
}
\description{
\packageDescription{npiv}
}
\details{
\packageDESCRIPTION{npiv}
\packageIndices{npiv}
This package provides a function with a simple interface for conducting nonparametric instrumental variable (IV) estimation, \code{npiv(...)} given a dependent variable vector \code{Y}, matrix of endogenous predictors \code{X}, and matrix of instruments \code{W}.
}
\author{
\packageAuthor{npiv}

Maintainer: Jeffrey S. Racine <racinej@mcmaster.ca> \packageMaintainer{npiv}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
See \code{\link{npiv}}
}
